requirements.txt
jax==0.4.30
jaxlib==0.4.30
flax==0.10.2
optax==0.2.4
chex==0.1.88
distrax==0.1.5
ml_collections==1.0.0
jaxopt==0.8.3
hydra-core==1.3.2
omegaconf==2.3.0
gym==0.26.2
gymnasium==1.0.0
smac==1.0.0
id-mava==0.2.0
jaxmarl==0.0.7
matplotlib==3.7.5
pandas==2.2.3
scipy==1.12.0
tensorboard==2.17.0